requirements.txt
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pyzmq==22.0.2
websockets==8.1
